==> Coursera-Blight/prepare/7-clustering.ipynb <==
# Basic
from datetime import datetime, timedelta
import time

# Data Analysis Specific
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from dateutil.parser import parse as date_parser
from geopy.geocoders import Nominatim

# Marchine Learning Specific
from sklearn.kernel_ridge import KernelRidge
from sklearn.ensemble import RandomForestRegressor, GradientBoostingRegressor
from sklearn.svm import SVR
from sklearn.neighbors import KNeighborsRegressor
from sklearn.grid_search import GridSearchCV, RandomizedSearchCV

# IPython magic
%matplotlib inline
%load_ext autoreload
%autoreload 2

DATA_PATH = './data/'
DETROIT_LAT = (42.252, 42.452)
DETROIT_LNG = (-83.295, -82.895)
SIZE_RATIO = 1000
DETROIT_WIDTH = 200
DETROIT_HEIGHT = 400
print DETROIT_WIDTH, DETROIT_HEIGHT, float(DETROIT_WIDTH)/DETROIT_HEIGHT

## The naive (grid) building method can not get good result, partially due to the naive method didn't consider the space relationship. 

## The grid method can be used if using convolutional neural network (just an assumption). For time limit, I didn't explore this idea. I might make this trial later.

## Now, it may be a good idea to use clustering (in fact this is the first idea I got), on the space only, to capture the space relationships.

## There are many clustering algos, I will first try kmeans, for simpleness.

clean_permit = pd.read_csv('clean/permit.csv',index_col=0)
clean_violation = pd.read_csv('clean/violation.csv',index_col=0)
clean_crime = pd.read_csv('clean/crime.csv',index_col=0)
clean_311 = pd.read_csv('clean/311.csv',index_col=0)

clean_crime.date.isnull().sum()

clean_permit.date.isnull().sum()

clean_permit[~clean_permit['turst']]

clean_violation.date.isnull().sum()

clean_311.date.isnull().sum()

len(clean_violation.date.dropna().sort_values())

def dropna(df):
    df = df.dropna()
    if hasattr(df,"turst"):
        df = df[df.turst]
    return df

addr = dropna(clean_violation)[['lat','lng']]

from sklearn.preprocessing import MinMaxScaler
class Scaler(MinMaxScaler):
    def transform(self,X):
        return pd.DataFrame(super(Scaler,self).transform(X),index=X.index,columns=X.columns)
scaler = Scaler().fit(addr)

addr = scaler.transform(addr)

# Give up using sklearn... too slow. Use Graphlab Create instead.

from sklearn.cluster import KMeans
cluster = KMeans(n_clusters=20000,max_iter=300)
cluster.fit(addr)

plt.scatter(*zip(*cluster.cluster_centers_))

export = pd.DataFrame(cluster.cluster_centers_,oolumns=addr.columns)

export.to_csv("center.csv")
