==> Coursera-Blight/prepare/11_build_features_without_time.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from sklearn.neighbors import NearestNeighbors
from sklearn.ensemble import GradientBoostingClassifier,GradientBoostingRegressor
from sklearn.cross_validation import cross_val_score
from sklearn.metrics import make_scorer

%matplotlib inline

clean_permit = pd.read_pickle('clean/permit.pickle')
clean_violation = pd.read_pickle('clean/violation.pickle') 
clean_crime = pd.read_pickle('clean/crime.pickle')
clean_311 = pd.read_pickle('clean/311.pickle')
clusters = pd.read_csv('clean/buildings.csv')[['lat','lng']]
def cut_date(df):
    return df[df.date.map(lambda x: x.year)>=2005]
clean_permit, clean_violation, clean_crime, clean_311 = map(cut_date,[clean_permit, clean_violation, clean_crime, clean_311])
print len(clusters)

pd.read_csv('clean/buildings.csv')[['lat','lng']]

nbrs = NearestNeighbors(n_neighbors=1).fit(clusters)

def find_nearest(df):
    distances, indices = nbrs.kneighbors(df[['lat','lng']])
    df['neighbor'] = indices
    return df
clean_permit, clean_violation, clean_crime, clean_311 = map(find_nearest,[clean_permit, clean_violation, clean_crime, clean_311])

clusters['judge'] = clean_violation.groupby('neighbor').apply(lambda g: g['JudgeAmt'].mean())
clusters['violation'] = clean_violation.groupby('neighbor').apply(lambda g: g['JudgeAmt'].count())

clusters['permit'] = clean_permit.groupby('neighbor').apply(lambda g: g['addr'].count())
clusters['permit'] = clusters['permit'].fillna(0)

clusters = pd.concat([clusters,pd.get_dummies(clean_crime['Category'],prefix='crime').groupby(clean_crime['neighbor']).sum()],axis=1)
clusters = clusters.fillna(0)

clusters = pd.concat([clusters,pd.get_dummies(clean_311['Category'],prefix='311').groupby(clean_311['neighbor']).sum()],axis=1)
clusters = clusters.fillna(0)

clusters['crime'] = clean_crime.groupby('neighbor').apply(lambda g: g['Category'].count())
clusters['crime'] = clusters['crime'].fillna(0)

clusters

# clusters[clusters['violation']>1000]=1000

def balanced_dtrain(raw, clf=True, sample=True, target='permit'):
    non_zero = raw[raw.sum(axis=1)!=0].copy()
    if clf:
        non_zero[target] = non_zero[target]>0
    positive = non_zero[non_zero[target]>0]
    negative = non_zero[non_zero[target]==0]
    if sample:
        negative = negative.sample(len(positive))
    return pd.concat([negative,positive]).sort_index()

dataset = balanced_dtrain(clusters).drop(['lat','lng'],axis=1)
X = dataset.drop('permit',axis=1)
y = dataset['permit']

clf = GradientBoostingClassifier(n_estimators=1000,max_depth=1)
cross_val_score(clf, X, y, cv=5,scoring="roc_auc").mean()

clf = GradientBoostingClassifier(n_estimators=500,max_depth=3)
cross_val_score(clf, X, y, cv=5,scoring="roc_auc").mean()

dataset = balanced_dtrain(clusters,clf=False).drop(['lat','lng'],axis=1)
X = dataset.drop('permit',axis=1)
y = dataset['permit']
scorer = make_scorer(lambda ground_truth, predictions: ((predictions>=1)==(ground_truth>=1)).mean())

rgr = GradientBoostingRegressor(n_estimators=500,max_depth=3)
cross_val_score(rgr, X, y, cv=5,scoring=scorer).mean()

rgr = GradientBoostingRegressor(n_estimators=800,max_depth=3)
cross_val_score(rgr, X, y, cv=5,scoring=scorer).mean()

rgr = GradientBoostingRegressor(n_estimators=800,max_depth=2)
cross_val_score(rgr, X, y, cv=5,scoring=scorer).mean()
