==> Coursera-Blight/prepare/12_addr_reset.ipynb <==
import pandas as pd

clean_permit = pd.read_pickle('clean/permit.pickle')
clean_violation = pd.read_pickle('clean/violation.pickle') 
clean_crime = pd.read_pickle('clean/crime.pickle')
clean_311 = pd.read_pickle('clean/311.pickle')
def cut_date(df):
    return df[df.date.map(lambda x: x.year)>=2005]
clean_permit, clean_violation, clean_crime, clean_311 = map(cut_date,[clean_permit, clean_violation, clean_crime, clean_311])

# thanks to Linghao Zhang zlhdnc1994@gmail.com
import string
import re

def clean_addr(s):
    s = s.lower()
    s = s.translate(string.maketrans('', ''), string.punctuation)
    s = re.sub(r'\bstreet\b', r'st', s)
    s = re.sub(r'\broad\b', r'rd', s)
    s = re.sub(r'\bavenue\b', r'av', s)
    s = re.sub(r'\bdrive\b', r'dr', s)
    s = re.sub(r'\bboulevard\b', r'bd', s)
    s = re.sub(r'\bblvd\b', r'bd', s)
    s = re.sub(r'\bnorth\b', r'n', s)
    s = re.sub(r'\bsouth\b', r's', s)
    s = re.sub(r'\bwest\b', r'w', s)
    s = re.sub(r'\beast\b', r'e', s)
    s = re.sub(r' +', r' ', s)
    p = s.find('detroit mi')
    if p != -1:
        s = s[:p]
    p = s.find('detroitmi')
    if p != -1:
        s = s[:p]        
    s = s.strip()
    return s

clean_violation.addr = clean_violation.addr.map(clean_addr)
clean_permit.addr = clean_permit.addr.map(clean_addr)

clean_permit.groupby(['addr','lat','lng']).count()

clean_permit.drop('date',axis=1).groupby(['addr']).apply(lambda g: len(g.apply(lambda g: (g['lat'],g['lng']),axis=1).unique()))
_[_!=1]

idx = clean_permit[clean_permit.addr=='3716 benson'].index
clean_permit.loc[idx,'lat'] = 0.529960
clean_permit.loc[idx,'lng'] = 0.709746

idx = clean_permit[clean_permit.addr=='4893 grayton'].index
clean_permit.loc[idx,'lat'] = 0.767547
clean_permit.loc[idx,'lng'] = 0.945206

clean_permit

len(clean_permit.addr.unique())

clean_violation.groupby(['addr','lat','lng']).count()

buildings = pd.concat([clean_permit[['addr','lat','lng']],clean_violation[['addr','lat','lng']]]).drop_duplicates("addr",keep="first").reset_index().drop('index',axis=1)

buildings.to_csv('clean/buildings.csv')

98674

buildings
